fix: Retain target parkrun in state until form is resubmitted
There is a bug where the target parkrun referenced is updated
when a new target parkrun is selected but before the form is
resubmitted. This is confusing and results in incorrect
information, e.g., 'Estimated time at Bushy Park' when actually
the estimated time shown is for the original target parkrun
selected.